% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_query_code.R
\name{seq_query}
\alias{seq_query}
\title{query function}
\usage{
seq_query(query_chr, query_pos_start, query_pos_end, proc_tgf_path,
  proc_chr_path, codon_dictionary_path)
}
\arguments{
\item{query_chr}{a string indicating the chromosome you want to query}

\item{query_pos_start}{an integer indicating the starting position you are intersted}

\item{query_pos_end}{an integer indicating the ending position you are intersted}

\item{proc_tgf_path}{a string indicating the path to the processed tgf file (provided by the package,"tgf_df_CDS.Rds")}

\item{proc_chr_path}{a string indicating the path to the directory of chromose fasta files (provided by the package,"chr*_seq.Rds")}

\item{codon_dictionary_path}{a string indicating the path to the codon directory (provided by the package, "codon_dictionary.Rds")}
}
\description{
This function allows you to query a chormosome and a position
}
\examples{
getit = seq_query(query_chr = "chr1",
                  query_pos_start = 70000,
                  query_pos_end = 70000,
                  proc_tgf_path = "/data/ginny/GENCODE",
                  proc_chr_path = "/data/ginny/GENCODE/chromosomes",
                  codon_dictionary_path = "/data/ginny/GENCODE")
}
